gear: replace hark types from landscape-apps
